Draft: Update RamCentralRepository with new search functionality

See merge request fsc-clam/fsc-wheres-what!41

+++ app/src/main/java/io/gitlab/fsc_clam/fscwhereswhat/model/database/DBEvent.kt
@@ -39,7 +39,7 @@ data class DBEvent(
 	val description: String,
 	val instructions: String?,
 	val locationName: String,
-	val locationId: Int,
+	val locationId: Long,
 	val hasRSVP: Boolean,
 	val url: String
 )
